I think I've finally done it, I have figured out a way to insert data from textboxes and insert them into SQL database.
Its unfortunate that the solution was so easy and right nuder my nose the entire time. The API is still not involved in the posting procedure, which needs to be fixed, and currently I have SQL and stored procedure code sitting in MVVM, and that is way to close to the UI for comfort. Also I have to rememebr to do some SQLinjection research to make sure that I am not suceptible to that. Which I am fairly certian I am currently.